Major thrashup to add map projections to CAP plots.  Many files have changed, a couple have been added.
Also fixed a zoom bug that's been there since the beginning (zoom on an
area, get something totally different -- the code which tweaks the display
area to match the window aspect ratio managed to hide this one for a
while).  When drawing a box for zooming, the box is now constrained to fit
the shape of the window, so the user gets a better idea of what the final
view will be.


git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/zebra/source@4828 3f9b6c58-9310-455a-aec6-8b30269ab333