Additions for the new way of life in S-Pol.  We now have a "polar" data chunk class for vector data in polar coordinates.  DCC_Polar also allows an internal "get beam" function so that an entire sweep need not be stored in memory at once.
We can now read sweepfiles.  Currently only polar and location datachunks
can be returned, though one can see applications for grid/nspace orgs as
well.  The notion of a "sample" in sweepfiles is one sweep; the main reason
for this is so that the radar altitude code in gp still has a chance of
working.


git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/zebra/source@6448 3f9b6c58-9310-455a-aec6-8b30269ab333