Added plots from polar data.
The new "polar" representation rasterizes a sweep in the usual way, providing
for thresholding and a couple of other things.  ga_GetGrid will also use the
rasterizing code to create regular grids from polar data, allowing it to be
contoured.

Also carved a lot of the low-level image code out of RasterPlot.c and created
RasterImage.c for it.  Now code dealing with images need not know about
shared memory, ximages, and the like.


git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/zebra/source@6449 3f9b6c58-9310-455a-aec6-8b30269ab333