Added -dsonly as a synonym for -ds.  Added -dshost to specify a DS_DAEMON_HOST in lieu of setting the environment variable; this is useful when running zstart through an xterm, i.e., "xterm -e zstart -dshost blah -dsonly".
The ssh startup on DS_DAEMON_HOST is now backgrounded, since ssh will
not exit until the child processes of the remote shell are finished, and
in this case that means until the remote Zebra is shut down.

Since DS_DAEMON_HOST remote startup now gets backgrounded, a msg_ping
test is now used later in the script to wait for the remote DS_Daemon
to be ready.  The wait may be need to be lengthened...


git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/zebra/source@7460 3f9b6c58-9310-455a-aec6-8b30269ab333